Refactoring User Actions into the User Action Processor
