+++ app/main/views/root.py
@@ -10,7 +10,6 @@ from app.authentication.invalid_token_exception import InvalidTokenException
 from app.main import errors
 from app.utilities.factory import factory
 from app.metadata.metadata_store import MetaDataStore
-from app.routing.routing_engine import RoutingEngine
 import logging
 
 
@@ -58,11 +57,8 @@ def login():
         # load the navigation history
         navigation_history = factory.create("navigation-history")
 
-        # Create the routing engine
-        routing_engine = RoutingEngine(schema)
-
         # create the navigator
-        navigator = Navigator(schema, metadata, navigation_history, routing_engine)
+        navigator = Navigator(schema, metadata, navigation_history)
 
         # get the current location of the user
         current_location = navigator.get_current_location()
